Add requested changes - Add initital tubes cost - Remove unnecessary console logs - Simplify solution to one file
